perf(gc): shrink RC header 64->56 B (P2 memory)
The RC/GC object header was 64 B (ref_count size_t + gc_flags + borrow_count +
gc_mark enum + 4 GC-list ptrs + 2 fn ptrs). Repack the leading fields into one
8-byte word: ref_count size_t->uint32_t (4 billion refs >> addressable memory),
gc_mark __yo_gc_mark_t(4)->uint8_t (values 0..4), keeping gc_flags(1)+borrow(2).
Header 64->56 B, so ~8 B off every RC object.

The win is size-class-dependent (macOS 16-byte quanta): objects that cross a
class boundary save 16 B. Measured on the yo-self self-compile heap profile
(macOS heap): ArrayList 88->80 B moves the 96 B class (26 M objects) to the 80 B
class, and EvalValue 104->96 likewise — total live heap 6.14 GB -> 5.72 GB
(-0.43 GB / 7%) at the mid-compile sampling point; RSS 5.8 -> 5.4 GB.

The atomic-RC path reinterprets &ref_count as _Atomic size_t*, so all 5 atomic
casts (incr/decr add/sub + typeid load in rc-fns.ts) are narrowed to
_Atomic uint32_t* to match the field width (lock-free on arm64); both header
variants (cyclic + lightweight) use uint32_t ref_count so the cast width is
uniform. GC_DEBUG %zu prints cast to (size_t). Mirrored 1-to-1 in the yo-self
emitter.

Validated: corpus 96/96, check ./std 152/152 (TS + yo-self binary), cycle_collector
16/16, arc/atomic_object/imm_threading/thread/thread_safety, closure_capture_rc_leak,
continue_rc_cleanup, algebraic_effects 72, option_result_combinators 54.

+++ src/codegen/exprs/rc-fns.ts
@@ -546,7 +546,7 @@ export function generateRcCall(
   if (isRcType(argType)) {
     if (isAtomicReferenceStructType(argType)) {
       // Atomic objects need atomic_load to avoid data races across threads
-      return `atomic_load_explicit((_Atomic size_t*)&((__yo_ref_header_t*)(${argCode}))->ref_count, memory_order_acquire)`;
+      return `atomic_load_explicit((_Atomic uint32_t*)&((__yo_ref_header_t*)(${argCode}))->ref_count, memory_order_acquire)`;
     }
     return `((__yo_ref_header_t*)(${argCode}))->ref_count`;
   } else {
